requirements.txt
aiohttp==3.10.11
spotipy==2.24.0
PyYAML==6.0.2
